Updated autonomous classes
Class 1 - AutoBlueRange
Updated the class to better match the updates made in AutoRedRange.
This includes the lower wait times to reduce time taken, gyro range
alignments, movement failsafe in case of realignment to before beacon
(driveAlongWallToBeaconOrForUnits).

Class 2 - AutoRedRange
Added turnParallelToWallWithGyro in order to speed up wall
stabilization. Coasting removed in order to prevent the
turning/crashing issue.

Class 3 - MyOpMode
Set found angles for lookUpTurningAngleForRangeDiff to the values found
from testing. Fixed a bug that involved the brokenness of the range
sensor not being checked in the turnParallelToWallWithGyro method.
Added a maximum number of recursive runs of turnParallelToWallWithGyro.
Added driveAlongWallToBeaconOrForUnits, which is designed to be a
failsafe in the event that the robot turns too far backwards after the
beacon is “found”. Changed turnParallelToWall to
turnParallelToWallWithGyro for the main range correction while driving
along the wall methods.

Class 4 - TurnToWallTest
Changed speed to reflect the current autonomous