League Meet 3 Commit
Class 1 - AutoBlueRange
Added voltage scaling for shooter
Increased initial movement by 100 encoder ticks.
Increased second encoder movement by 1000 ticks.
gyroArcTurnRight turns “yawDiff-13.0” degrees in order to better align
with the wall.
turnParallelToWallWithGyro speeds decreased to increase accuracy.
stabilizeAlongWallWithRange methods have target distance increased by 1
and threshold decreased by 1, this is a result of the robot driving too
close to the wall and not adjusting soon enough before it hits the wall.
Changed movement after first beacon press to be a
stabilizeAlongWallWithRange method instead of moveWithEncoders,
increases accuracy.
Added another case for pressing the second beacon that adds the
failsafe driveAlongWallToBeaconOrForUnits method.

Class 2 - AutoRedRange
Lowered battery threshold to reflect a more accurate representation of
what the running battery voltage is at.
Increased movement after shooting by 200 encoder ticks
Decreased the gyroArcTurnRight by 7 degrees in order to better align
with the wall.
Slightly decreased turnParallelToWallWithGyro speeds in hopes of
improving accuracy.

Class 3 - MoveAlongWallToBeaconTest
stabilizeAlongWallWithRangeToBeacon changed to
stabilizeAlongWallWithRangeForEncoderDist for code testing purposes.

Class 4 - MyAutonomous
initAngle variable initialized in MyAutonomous. See “MyOpMode” changes.

Class 5 - MyOpMode
New global variable added to “MyOpMode” called initAngle. This angle
reflects the initial heading of the robot before autonomous alignment,
and represents the heading that would make the robot parallel with the
beacons.
Added maximum turntime for the gyroTurn methods of 1 second in order to
reduce problems in autonomous with gyro turns taking too long.
Power now scales from 100% of speed variable to 83% of it in gyroTurns
instead of 100% and 50%. This change was made due to turns not being
completed.
Added a case 0 to lookUpTurningAngleForRangeDiff in case the difference
happens to be 0.
turnParallelToWallWithGyroSimple created in order to allow only one run
of the method, as well as only run the method if it’s necessary when
the robot is too close to the wall or too far from the wall.
Added stabilize failsafe in case the range sensor returns a much lower
value than expected. This is done by comparing the gyro heading to the
calculated heading via range sensors. If the calculated heading is too
different from the gyro heading, the gyro heading is used instead of
the range heading.
Increased the maximum drifting angle to be a difference of 5.0 range
units.
Power of drifting inwards from case 00 increased.