Saturday testing commit
Gave all movement methods a simple drift offset multiplication value.
Currently the left side value is 0.93 and the right side value is 1.0

AutoRedSimple and AutoBlueSimple have been created. These are expected
to be the primary autonomouses used from now on. There is no direct
stopping like there was before, where the motors reversed in order to
stop the robot. Instead, coasting is implemented. Based on initial
testing, coasting increases the accuracy of the autonomous. A primary
goal of the “Simple” autonomous programs is to no longer rely on the
range sensors for much at all. Instead all stabilization will be
gyro-based. The only use I currently plan to have for the range sensors
is to sense the robot’s distance from the wall and push out the beacon
pusher an appropriate distance. The only correction I currently have
planned is a gyro turn back to the initial heading after the first
beacon push. In general, the movements of the Simple classes are
somewhat similar to the movements of the old autonomous programs,
except the robot presses the near beacon before the further away beacon.

moveWithEncodersCoast was created in order to allow for the coast
movements that are used in the “Simple” autonomous programs. The power
scales from initial power to 1/3rd power throughout the movement in
order to increase precision of movement.

TeleOp has had its threshold for trigger presses increased to 0.6
instead of 0.4. This is the result of testing that indicated that the
“ghost value” sent from the controller is a value of 0.5. The hope is
that this change will prevent the problems of the triggers being
pressed throughout the match.