Added base of macros and rollers
MyOpMode
The essential roller code is complete. The movement methods are not
tested, but they are theoretically close.

TeleOp
The base code for the macros that drive along the wall to the beacons,
as well as the turn to other wall and press beacon are completed. Some
of it needs to be completed.

SpinnerTest
The estimated encoderTicks max speed value is now 3670, which in
testing seemed like the right spinner speed value.